Major changes to Farey polynomials (fix #7)
Instead of taking α and β values, Farey polynomial functions
now take trX, trY, and trXY. This is the first step to allowing
computation of substitution slices:- the (p,q) Riley slice should
be the intersection of the exterior of the closures of the sets

W_r/s(X^m, Y^n)

as m runs over 1,...,p and n runs over 1,...,q.

Also a lot of tests have been added.