Try to work around mpmath pickling error so atom.py works off linux
Due to https://github.com/mpmath/mpmath/issues/380 we can't pickle the
group G in atom.py. This means that unless there is some other way of sharing
the object to subprocesses we have to regenerate the group in all of them. On
linux we can fork to get the subprocess, so G is automatically shared. On Win32
or OS X this is not possible, so check if G is defined and if not regenerate it.

Cf. #19